# protos/task.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "protos/task.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/tasks": {
      "get": {
        "operationId": "ListTask",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1TaskCollection"
            }
          }
        },
        "parameters": [
          {
            "name": "fields",
            "description": "reserved 1 to 5;\nreserved \"fields\", \"expands\", \"page\", \"limit\", \"offset\",\"sort\", \"filter\",\"count\",\"sum\",\"context\";.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "filter",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "count",
            "in": "query",
            "required": false,
            "type": "boolean",
            "format": "boolean"
          },
          {
            "name": "sum",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "context",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "not_completed",
            "in": "query",
            "required": false,
            "type": "boolean",
            "format": "boolean"
          }
        ],
        "tags": [
          "TaskService"
        ]
      },
      "post": {
        "operationId": "CreateTask",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1TaskEntity"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1Task"
            }
          }
        ],
        "tags": [
          "TaskService"
        ]
      }
    },
    "/tasks/{id}": {
      "get": {
        "operationId": "GetTask",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1TaskEntity"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "TaskService"
        ]
      },
      "delete": {
        "operationId": "DeleteTask",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteTaskResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "TaskService"
        ]
      },
      "put": {
        "operationId": "UpdateTask",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1TaskEntity"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1Task"
            }
          }
        ],
        "tags": [
          "TaskService"
        ]
      }
    },
    "/tasks/{id}:complete": {
      "post": {
        "operationId": "CompleteTask",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1TaskEntity"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "TaskService"
        ]
      }
    }
  },
  "definitions": {
    "LinkMethod": {
      "type": "string",
      "enum": [
        "PUT",
        "GET",
        "HEAD",
        "POST",
        "PATCH",
        "DELETE",
        "TRACE",
        "OPTIONS"
      ],
      "default": "PUT",
      "title": "https://tools.ietf.org/html/rfc7231#section-4"
    },
    "v1Complete": {
      "type": "string",
      "enum": [
        "UNKNOWN",
        "RUNNING",
        "COMPLETED"
      ],
      "default": "UNKNOWN"
    },
    "v1DeleteTaskResponse": {
      "type": "object"
    },
    "v1Link": {
      "type": "object",
      "properties": {
        "rel": {
          "type": "string"
        },
        "method": {
          "$ref": "#/definitions/LinkMethod"
        },
        "href": {
          "type": "string"
        },
        "type": {
          "type": "string"
        }
      },
      "title": "/-------------------------------------------------------------------\n/ List of official link rels:\n/ http://www.iana.org/assignments/link-relations/link-relations.xhtml\n/-------------------------------------------------------------------"
    },
    "v1Task": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "completed": {
          "$ref": "#/definitions/v1Complete"
        },
        "created_at": {
          "type": "string",
          "format": "date-time"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "v1TaskCollection": {
      "type": "object",
      "properties": {
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1TaskEntity"
          }
        },
        "links": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1Link"
          }
        }
      }
    },
    "v1TaskEntity": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/v1Task"
        },
        "links": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1Link"
          }
        }
      }
    }
  }
}